Add files via upload
The data provided above represents a sample of 250 sentence pairs extracted from each literary work. The remainder of the content will be uploaded in subsequent phases. Please note that the dataset is currently under active development and expansion. We anticipate releasing additional content and enhancements to the corpus in the near future. We appreciate your interest in this resource and encourage you to check back for updates as we continue to build and refine this comprehensive collection of annotated classical Chinese literature.